requirements.txt
requests==2.31.0
schedule==1.1.0
pyTelegramBotAPI==4.9.0
python-decouple==3.7
pycryptodome==3.18.0